requirements.txt
streamlit
torch
transformers
openai==0.28
Pillow
python-dotenv
